Add first part of sensitivity analysis for epistemic uncertainties
Add partial implementation of sensitivity analysis for the conclusion of the cost-benefit analysis considering epistemic uncertainties. The first part includes the mapping of the epistemic uncertainty space and the calculation of the break-even values for the uncertain parameters for which the costs of risk-reducing measures balance the benefits. Calculation and plotting of sensitivity indices and some cleaning up of function interfaces remain.